Corrección de errores en los tipos de datos

+++ src/main/java/co/edu/javeriana/jpa_example2/model/Partida.java
@@ -41,7 +41,7 @@ public class Partida {
         return id;
     }
 
-    public int getTiempoLimite() {
+    public Long getTiempoLimite() {
         return tiempoLimite;
     }
